Separate most components into containers
