+++ dawn/renderer/components/Editor.js
@@ -1,19 +1,4 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import {
-  editorUpdate,
-} from '../actions/EditorActions.js';
-import {
-  changeTheme,
-  increaseFontsize,
-  decreaseFontsize,
-} from '../actions/SettingsActions.js';
-import {
-  showConsole,
-  hideConsole,
-  clearConsole,
-} from '../actions/ConsoleActions';
-import { addAsyncAlert } from '../actions/AlertActions';
 import EditorToolbar from './EditorToolbar';
 import ConsoleOutput from './ConsoleOutput';
 import Ansible from '../utils/Ansible';
@@ -37,7 +22,7 @@ import 'brace/theme/solarized_dark';
 import 'brace/theme/solarized_light';
 import 'brace/theme/terminal';
 
-class EditorComponent extends React.Component {
+class Editor extends React.Component {
   constructor(props) {
     super(props);
     this.themes = [
@@ -222,7 +207,7 @@ class EditorComponent extends React.Component {
   }
 }
 
-EditorComponent.propTypes = {
+Editor.propTypes = {
   editorCode: React.PropTypes.string,
   editorTheme: React.PropTypes.string,
   filepath: React.PropTypes.string,
@@ -242,42 +227,4 @@ EditorComponent.propTypes = {
   runtimeStatus: React.PropTypes.bool,
 };
 
-const mapStateToProps = (state) => ({
-  editorCode: state.editor.editorCode,
-  editorTheme: state.settings.editorTheme,
-  filepath: state.editor.filepath,
-  fontSize: state.settings.fontSize,
-  latestSaveCode: state.editor.latestSaveCode,
-  showConsole: state.studentConsole.showConsole,
-  consoleData: state.studentConsole.consoleData,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  onAlertAdd: (heading, message) => {
-    dispatch(addAsyncAlert(heading, message));
-  },
-  onEditorUpdate: (newVal) => {
-    dispatch(editorUpdate(newVal));
-  },
-  onChangeTheme: (theme) => {
-    dispatch(changeTheme(theme));
-  },
-  onIncreaseFontsize: () => {
-    dispatch(increaseFontsize());
-  },
-  onDecreaseFontsize: () => {
-    dispatch(decreaseFontsize());
-  },
-  onShowConsole: () => {
-    dispatch(showConsole());
-  },
-  onHideConsole: () => {
-    dispatch(hideConsole());
-  },
-  onClearConsole: () => {
-    dispatch(clearConsole());
-  },
-});
-
-const Editor = connect(mapStateToProps, mapDispatchToProps)(EditorComponent);
 export default Editor;
