+++ dawn/renderer/components/FinalCompPeripheralListContainer.js
@@ -1,0 +1,10 @@
+import { connect } from 'react-redux';
+import FinalCompPeripheralList from './FinalCompPeripheralList';
+
+const mapStateToProps = (state) => ({
+  peripherals: state.peripherals,
+});
+
+const FinalCompPeripheralListContainer = connect(mapStateToProps)(FinalCompPeripheralList);
+
+export default FinalCompPeripheralListContainer;
