+++ dawn/renderer/components/GamepadsContainer.js
@@ -1,0 +1,10 @@
+import { connect } from 'react-redux';
+import Gamepads from './Gamepads';
+
+const mapStateToProps = (state) => ({
+  gamepads: state.gamepads.gamepads,
+});
+
+const GamepadsContainer = connect(mapStateToProps)(Gamepads);
+
+export default GamepadsContainer;
