+++ dawn/renderer/components/peripherals/GenericPeripheral.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import NameEdit from '../NameEdit';
+import NameEditContainer from '../NameEditContainer';
 
 /**
  * A generic peripheral, used when the peripheralType is unknown.
@@ -8,7 +8,7 @@ const GenericPeripheral = (props) => (
   <div style={{ overflow: 'auto' }}>
     <div style={{ overflow: 'auto', width: '100' }}>
       <h4 style={{ float: 'left' }}>
-        <NameEdit name={props.name} id={props.id} />
+        <NameEditContainer name={props.name} id={props.id} />
         <small>{props.peripheralType}</small>
       </h4>
       <h4 style={{ float: 'right' }}>
