+++ dawn/renderer/components/peripherals/ScalarSensor.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import NameEdit from '../NameEdit';
+import NameEditContainer from '../NameEditContainer';
 import numeral from 'numeral';
 
 const ScalarSensor = (props) => (
   <div style={{ overflow: 'auto' }}>
     <div style={{ overflow: 'auto', width: '100%' }}>
       <h4 style={{ float: 'left' }}>
-        <NameEdit name={props.name} id={props.id} />
+        <NameEditContainer name={props.name} id={props.id} />
         <small> {props.peripheralType} </small>
       </h4>
       <h4 style={{ float: 'right' }}> {numeral(props.value).format('0.00')} </h4>
